pyproject.toml
[tool.poetry]
name = "fsdsdf"
version = "0.1.0"
description = ""
authors = ["Foshati <amirrezafoshati@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
django = "^5.0.6"
psycopg2 = "^2.9.9"
django-cities-light = "^3.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
